e2e tests for completion command (#313)
* e2e tests for completion cmd

* change

* change

---------
